Struct pe_util::PE

#[repr(transparent)]
pub struct PE<'a> { /* private fields */ }

A pointer sized type that allows the user to read a buffer in memory as a Windows PE. Currently only supports 32-bit and 64-bit PEs on x86 architectures, but, I plan on supporting more architectures in the future. This type is indifferent to 32 or 64-bit architecture, as well as whether the file has been mapped into memory for execution, or if the file is still in its on disk stat.

Example

fn example(slice: &[u8]) {
    let pe = PE::from_slice(slice).expect("Could not validate that slice is a valid PE file.");
    let exports = pe.get_exports().unwrap_or_default();

    for export in exports {
        let ord = pe.get_function_ordinal(export.as_bytes());
        println!("Export: {export} Ordinal: {ord}");
    }

    let res = pe.get_pe_resource(10, 100).expect("Could not find PE resource");

    println!("{}", res.len());
}

Implementations

impl<'a> PE<'a>

pub fn from_slice(slice: &'a [u8]) -> Result<Self>

Returns a PE from a slice that starts with a valid PE file. The returned PE shares the lifetime of the slice. Returns an Error with ErrorKind::InvalidData if the pe cannot be validated by magic bytes and header signature values.

pub unsafe fn from_slice_unchecked(slice: &'a [u8]) -> Self

Returns a PE from a slice that starts with a valid PE file. The returned PE shares the lifetime of the slice. Does not validate that the slice is a valid PE file.

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the slice is a valid PE file.

pub unsafe fn from_slice_assume_mapped(slice: &'a [u8], is_mapped: bool) -> Self

Returns a PE from a slice that shares the lifetime of the slice. Assumes the mapped state of with the passed in is_mapped parameter, and does not validate that the slice is a valid PE file.

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the slice is a valid PE file. The caller must specify the correct mapped state for is_mapped.

pub unsafe fn from_ptr(ptr: *const u8) -> Result<Self>

Returns a PE from a *const u8 that points to the start of a valid PE file. There is no associated lifetime for the returned PE. Returns an Error with ErrorKind::InvalidData if the pe cannot be validated by magic bytes and header signature values.

Safety

pub unsafe fn from_ptr_unchecked(ptr: *const u8) -> Self

Returns a PE from a *const u8 that points to the start of a valid PE file. There is no associated lifetime for the returned PE. Does not do any validation of the data at the pointer.

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the pointer points to a valid PE file.

pub unsafe fn from_ptr_assume_mapped(slice: *const u8, is_mapped: bool) -> Self

Returns a PE from a *const u8 that points to the start of a valid PE file. There is no associated lifetime for the returned PE. Assumes the mapped state of with the passed in is_mapped parameter, and does not validate that the slice is a valid PE file.

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the slice is a valid PE file. The caller must specify the correct mapped state for is_mapped.

pub unsafe fn from_address(base_address: usize) -> Result<Self>

Returns a PE from a usize with the value of the address of a valid PE file. There is no associated lifetime for the returned PE. Returns an Error with ErrorKind::InvalidData if the pe cannot be validated by magic bytes and header signature values.

Safety

This method does not share lifetime with anything

impl PE<'_>

pub fn is_64bit(&self) -> bool

Returns true if the architecture for the PE is 64-bit, or false for any other architecture.

pub fn is_mapped(&self) -> bool

Returns true if the PE is mapped into memory, or false is in it’s “on disk” state.

pub fn validate(&self) -> bool

Checks that the memory pointed to by self.pointer is a valid PE file.

pub fn dos_headers(&self) -> &DosHeader<'_>

Returns the DosHeader structure, which allows you to inspect the PEs IMAGE_DOS_HEADER structure.

pub unsafe fn dos_headers_mut(&mut self) -> &mut DosHeader<'_>

Returns the DosHeader structure, which allows you to inspect and mutate the PEs IMAGE_DOS_HEADER structure.

Safety

The caller has to guarantee that they are upholding rusts mutability invariance with the original buffer that holds the PE. Mutating the DosHeader of a PE that has another mutable reference could result in undefined behaviour. Edit PE files your program doesn’t own at your own risk.

pub fn nt_headers(&self) -> &NtHeaders<'_>

Returns the NtHeaders structure, which allows you to inspect the PEs IMAGE_NT_HEADERS structure.

pub unsafe fn nt_headers_mut(&mut self) -> &mut NtHeaders<'_>

Returns the NtHeaders structure, which allows you to inspect and mutate the PEs IMAGE_NT_HEADERS structure.

Safety

The caller has to guarantee that they are upholding rusts mutability invariance with the original buffer that holds the PE. Mutating the NtHeaders of a PE that has another mutable reference could result in undefined behaviour. Edit PE files your program doesn’t own at your own risk.

pub fn section_headers(&self) -> &[IMAGE_SECTION_HEADER]

Returns the section headers for the PE file as a slice.

pub unsafe fn section_headers_mut(&mut self) -> &mut [IMAGE_SECTION_HEADER]

Returns the section headers for the PE file as a mutable slice.

Safety

The caller has to guarantee that they are upholding rusts mutability invariance with the original buffer that holds the PE. Mutating the IMAGE_SECTION_HEADER of a PE that has another mutable reference could result in undefined behaviour. Edit PE files your program doesn’t own at your own risk.

pub fn rva_to_foa(&self, rva: u32) -> Option<u32>

Takes the Relative Virtual Address and returns the File Offset Address. The return value is an offset.

pub fn get_export_address(&self, export: FunctionId<'_>) -> Option<usize>

Takes in a FunctionId and looks up the function address in the Export Directory by Name or Ordinal. Returns None if the specified function name or ordinal cannot be found.

pub fn get_export_rva(&self, export: FunctionId<'_>) -> Option<u32>

Takes in a FunctionId and looks up the function rva in the Export Directory by Name or Ordinal. Returns None if the specified function name or ordinal cannot be found.

pub fn get_exports(&self) -> Option<Vec<&str>>

Returns a Vec<&str> that contains the names all the functions that are exported by name for the PE. Returns None if it could not find the export directory.

pub fn get_export_name(&self, ordinal: u16) -> Option<String>

Returns the name of the function with the given ordinal value. Returns None if the Export directory cannot be found, or the provided ordinal is not found.

pub fn get_function_ordinal(&self, function_name: &[u8]) -> u16

Returns the ordinal of the function with the given name provided as a slice of bytes. Returns None if the Export directory cannot be found, or the provided ordinal is not found.

pub fn get_pe_resource( &self, category_id: u32, resource_id: u32 ) -> Option<&[u8]>

Looks up the Resource entry with the provided category and resource id. Currently, assumes the first directory in the language entries. Returns the resource as a &[u8], or None if the resource directory or resource could not be found.