Struct pe_util::PE

pub struct PE<'a, S> { /* private fields */ }

A pointer sized type that allows the user to read a buffer in memory as a Windows PE. Currently only supports 32-bit and 64-bit PEs on x86 architectures, but, I plan on supporting more architectures in the future. This type is indifferent to 32 or 64-bit architecture, as well as wether or not the file has been mapped into memory for execution, or if the file is still in it’s on disk stat.

Example

fn example(slice: &[u8]) {
    let pe = PE::from_slice(slice).expect("Could not validate that slice is a valid PE file.");
    let exports = pe.get_exports().unwrap_or_default();

    for export in exports {
        let ord = pe.get_function_ordinal(export.as_bytes());
        println!("Export: {export} Ordinal: {ord}");
    }

    let res = pe.get_pe_resource(10, 100).expect("Could not find PE resource");

    println!("{}", res.len());
}

Implementations

impl<'a> PE<'a, DosHeader>

pub fn from_slice(slice: &'a [u8]) -> Result<Self>

Returns a PE from a slice that starts with a valid PE file. The returned PE shares the lifetime of the slice. Returns an Error with ErrorKind::InvalidData if the pe cannot be validated by magic bytes and header signature values.

Arguments

slice: &'a [u8]

returns: Result<PE<DosHeader>, Error>

pub unsafe fn from_slice_unchecked(slice: &'a [u8]) -> Self

Returns a PE from a slice that starts with a valid PE file. The returned PE shares the lifetime of the slice. Does not validate that the slice is a valid PE file.

Arguments

slice: &'a [u8]

returns: PE<DosHeader

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the slice is a valid PE file.

pub unsafe fn from_slice_assume_mapped(slice: &'a [u8], is_mapped: bool) -> Self

Returns a PE from a slice that shares the lifetime of the slice. Assumes the mapped state of with the passed in is_mapped parameter, and does not validate that the slice is a valid PE file.

Arguments

slice: &'a [u8]

is_mapped: bool

returns: PE<DosHeader>

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the slice is a valid PE file.

impl PE<'_, DosHeader>

pub fn from_ptr(ptr: *const u8) -> Result<Self>

Returns a PE from a *const u8 that points to the start of a valid PE file. There is no associated lifetime for the returned PE. Returns an Error with ErrorKind::InvalidData if the pe cannot be validated by magic bytes and header signature values.

Arguments

ptr: *const u8

returns: Result<PE<DosHeader>, Error>

pub unsafe fn from_ptr_unchecked(ptr: *const u8) -> Self

Returns a PE from a *const u8 that points to the start of a valid PE file. There is no associated lifetime for the returned PE. Does not do any validation of the data at the pointer.

Arguments

ptr: *const u8

returns: DosHeader

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the pointer points to a valid PE file.

pub fn from_address(base_address: usize) -> Result<Self>

Returns a PE from a usize with the value of the address of a valid PE file. There is no associated lifetime for the returned PE. Returns an Error with ErrorKind::InvalidData if the pe cannot be validated by magic bytes and header signature values.

Arguments

base_address: usize

returns: Result<PE<DosHeader>, Error>

pub unsafe fn from_address_unchecked(base_address: usize) -> Self

Returns a PE from a usize with the value of the address of a PE file. There is no associated lifetime for the returned PE. Does not do any validation of the data at the address.

Arguments

base_address: usize

returns: PE<DosHeader>

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the address provided points to a valid PE file.

pub unsafe fn from_address_assume_mapped( base_address: usize, is_mapped: bool ) -> Self

Returns a PE from a usize with the value of the address of a PE file. There is no associated lifetime for the returned PE. Assumes the mapped state of with the passed in is_mapped parameter, and does not validate that the data at the address is a valid PE file.

Arguments

base_address: usize

is_mapped: bool

returns: PE<DosHeader>

Safety

This function does not check EncodedPointer compatability, mapped state, nor that the pointer points to a valid PE file.

pub fn validate(self) -> bool

Checks that the memory pointed to by self.pointer is a valid PE file.

returns: bool

pub fn dos_header(&self) -> &IMAGE_DOS_HEADER

Returns a reference to the start of the PE file as an IMAGE_DOS_HEADER.

returns: &'_ IMAGE_DOS_HEADER

pub fn dos_header_mut(&mut self) -> &mut IMAGE_DOS_HEADER

Returns a reference to the start of the PE file as an IMAGE_DOS_HEADER.

returns: &'_ IMAGE_DOS_HEADER

pub fn nt_headers(&self) -> PE<'_, NtHeaders>

Returns the NtHeaders variant of the PE structure.

returns: PE<NtHeaders>

pub fn section_headers(&self) -> &[IMAGE_SECTION_HEADER]

Returns the section headers for the PE file as a slice.

returns: &'_ mut [IMAGE_SECTION_HEADER]

pub fn section_headers_mut(&mut self) -> &mut [IMAGE_SECTION_HEADER]

Returns the section headers for the PE file as a mutable slice.

returns: &'_ mut [IMAGE_SECTION_HEADER]

pub fn rva_to_foa(self, rva: u32) -> Option<u32>

Takes the Relative Virtual Address and returns the File Offset Address. The return value is an offset.

Arguments

rva: u32

returns: Option<u32>

pub fn get_export_rva(self, export: FunctionId<'_>) -> Option<u32>

Takes in a FunctionId and looks up the function in the Export Directory by Name or Ordinal. Returns None if the specified function name or ordinal cannot be found.

Arguments

export: FunctionId

returns: Option<u32>

pub fn get_exports(&self) -> Option<Vec<&str>>

Returns a Vec<&str> that contains the names all the functions that are exported by name for the PE. Returns None if it could not find the export directory.

Arguments

export: FunctionId

returns: Option<Vec<&'_ str>>

pub fn get_export_name(self, ordinal: u16) -> Option<String>

Returns the name of the function with the given ordinal value. Returns None if the Export directory cannot be found, or the provided ordinal is not found.

Arguments

ordinal: u16

returns: Option<String>

pub fn get_function_ordinal(self, function_name: &[u8]) -> u16

Returns the ordinal of the function with the given name provided as a slice of bytes. Returns None if the Export directory cannot be found, or the provided ordinal is not found.

Arguments

function_name: &[u8]

returns: u16

pub fn get_pe_resource( &self, category_id: u32, resource_id: u32 ) -> Option<&[u8]>

Looks up the Resource entry with the provided category and resource id. Currently assumes the first directory in the language entries. Returns the resource as a &[u8], or None if the resource directory or resource could not be found.

Arguments

category_id: u32

resource_id: u32

returns: Option<&[u8]>

impl PE<'_, NtHeaders>

pub fn signature(&self) -> u32

pub fn file_header(&self) -> &IMAGE_FILE_HEADER

pub fn optional_header(&self) -> PE<'_, ImageOptionalHeader>

pub fn size_of(self) -> usize

impl PE<'_, NtHeaders>

pub fn set_signature(&mut self, value: u32)

pub fn file_header_mut(&mut self) -> &mut IMAGE_FILE_HEADER

impl PE<'_, ImageOptionalHeader>

pub fn magic(self) -> u16

pub fn major_linker_version(self) -> u8

pub fn minor_linker_version(self) -> u8

pub fn size_of_code(self) -> u32

pub fn size_of_initialized_data(self) -> u32

pub fn size_of_uninitialized_data(self) -> u32

pub fn address_of_entry_point(self) -> u32

pub fn base_of_code(self) -> u32

pub fn image_base(self) -> u64

pub fn section_alignment(self) -> u32

pub fn file_alignment(self) -> u32

pub fn major_operating_system_version(self) -> u16

pub fn minor_operating_system_version(self) -> u16

pub fn major_image_version(self) -> u16

pub fn minor_image_version(self) -> u16

pub fn major_subsystem_version(self) -> u16

pub fn minor_subsystem_version(self) -> u16

pub fn win32_version_value(self) -> u32

pub fn size_of_image(self) -> u32

pub fn size_of_headers(self) -> u32

pub fn check_sum(self) -> u32

pub fn subsystem(self) -> u16

pub fn dll_characteristics(self) -> u16

pub fn size_of_stack_reserve(self) -> u64

pub fn size_of_stack_commit(self) -> u64

pub fn size_of_heap_reserve(self) -> u64

pub fn size_of_heap_commit(self) -> u64

pub fn loader_flags(self) -> u32

pub fn number_of_rva_and_sizes(self) -> u32

pub fn data_directory(&self) -> &[IMAGE_DATA_DIRECTORY]

pub fn size_of(self) -> usize

impl PE<'_, ImageOptionalHeader>

pub fn set_magic(&mut self, value: u16)

pub fn set_major_linker_version(&mut self, value: u8)

pub fn set_minor_linker_version(&mut self, value: u8)

pub fn set_size_of_code(&mut self, value: u32)

pub fn set_size_of_initialized_data(&mut self, value: u32)

pub fn set_size_of_uninitialized_data(&mut self, value: u32)

pub fn set_address_of_entry_point(&mut self, value: u32)

pub fn set_base_of_code(&mut self, value: u32)

pub fn set_image_base(&mut self, value: u64)

pub fn set_section_alignment(&mut self, value: u32)

pub fn set_file_alignment(&mut self, value: u32)

pub fn set_major_operating_system_version(&mut self, value: u16)

pub fn set_minor_operating_system_version(&mut self, value: u16)

pub fn set_major_image_version(&mut self, value: u16)

pub fn set_minor_image_version(&mut self, value: u16)

pub fn set_major_subsystem_version(&mut self, value: u16)

pub fn set_minor_subsystem_version(&mut self, value: u16)

pub fn set_win32_version_value(&mut self, value: u32)

pub fn set_size_of_image(&mut self, value: u32)

pub fn set_size_of_headers(&mut self, value: u32)

pub fn set_check_sum(&mut self, value: u32)

pub fn set_subsystem(&mut self, value: u16)

pub fn set_dll_characteristics(&mut self, value: u16)

pub fn set_size_of_stack_reserve(&mut self, value: u64)

pub fn set_size_of_stack_commit(&mut self, value: u64)

pub fn set_size_of_heap_reserve(&mut self, value: u64)

pub fn set_size_of_heap_commit(&mut self, value: u64)

pub fn set_loader_flags(&mut self, value: u32)

pub fn set_number_of_rva_and_sizes(&mut self, value: u32)

pub fn data_directory_mut(&mut self) -> &mut [IMAGE_DATA_DIRECTORY]

impl<S> PE<'_, S>

pub fn base_address(self) -> usize

Returns the base address of the PE .

returns: usize

pub fn is_64bit(self) -> bool

Returns true if the architecture for the PE is 64-bit, or false for any other architecture.

returns: bool

pub fn is_mapped(self) -> bool

Returns true if the PE is mapped into memory, or false is in it’s “on disk” state.

returns: bool